change int -> size_t and add necessary bound checks and type casts

+++ src/timesync/nts_crypto.h
@@ -18,9 +18,9 @@ typedef struct AssociatedData {
  * RETURNS: the number of bytes in the ciphertext (< 0 indicates an error)
  */
 int NTS_encrypt(uint8_t *ctxt,
-                int ctxt_len,
+                size_t ctxt_len,
                 const uint8_t *ptxt,
-                int ptxt_len,
+                size_t ptxt_len,
                 const AssociatedData *info,
                 const NTS_AEADParam *aead,
                 const uint8_t *key);
@@ -35,9 +35,9 @@ int NTS_encrypt(uint8_t *ctxt,
  * RETURNS: the number of bytes in the decrypted plaintext (< 0 indicates an error)
  */
 int NTS_decrypt(uint8_t *ptxt,
-                int ptxt_len,
+                size_t ptxt_len,
                 const uint8_t *ctxt,
-                int ctxt_len,
+                size_t ctxt_len,
                 const AssociatedData *info,
                 const NTS_AEADParam *aead,
                 const uint8_t *key);
